backend/requirements.txt
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
mysqlclient==2.1.1
social-auth-app-django==4.0.0
Pillow==9.2.0
django-dbbackup==4.0.1
django-phonenumber-field==7.0.0
phonenumberslite==8.12.53
